Divided solution into two projects
