+++ alarm-mega/Debug/alarm-mega.hex
@@ -1,0 +1,37 @@
+:1000000071C000008EC000008CC000008AC00000DB
+:1000100088C0000086C0000084C0000082C00000CC
+:1000200080C000007EC000007CC000007AC00000DC
+:1000300078C0000076C0000074C0000072C00000EC
+:1000400070C000006EC000006CC000006AC00000FC
+:1000500068C0000066C0000064C0000062C000000C
+:1000600060C000005EC000005CC000005AC000001C
+:1000700058C0000056C0000054C0000052C000002C
+:1000800050C000004EC000004CC000004AC000003C
+:1000900048C0000046C0000044C0000042C000004C
+:1000A00040C000003EC000003CC000003AC000005C
+:1000B00038C0000036C0000034C0000032C000006C
+:1000C00030C000002EC000002CC000002AC000007C
+:1000D00028C0000026C0000024C0000022C000008C
+:1000E00020C0000011241FBECFEFD1E2DEBFCDBF84
+:1000F00000E00CBF12E0A0E0B2E0EEE2F2E000E0CF
+:100100000BBF02C007900D92AE30B107D9F722E0C5
+:10011000AEE0B2E001C01D92A431B207E1F702D017
+:1001200084C06ECFCF93DF93CDB7DEB76C970FB699
+:10013000F894DEBF0FBECDBF2EE0CE010196FC01CC
+:10014000322F11923A95E9F732E03C8347ED50E0C7
+:100150005A874987DE011F96FD0111922A95E9F71A
+:1001600021E02A8B2EED30E03A8F298FEEE0F2E08D
+:1001700093838283B183A08387E690E00DD080E0F3
+:1001800092E023D0FFE724E83EE1F15020403040E8
+:10019000E1F700C00000F3CF9093C5008093C40046
+:1001A00088E18093C1008EE08093C2000895909111
+:1001B000C00095FFFCCF8093C60008958091C000D9
+:1001C0008823E4F78091C60008950F931F93CF937F
+:1001D000DF93E0911002F0911102238121FF1BC0F7
+:1001E0008C01D0E0C0E0F80181918F016091100294
+:1001F00070911102DB011896ED91FC9119978823FB
+:1002000031F01995892B79F3DFEFCFEFECCF8AE04E
+:100210001995892B19F08FEF9FEF02C08D2F9C2F1E
+:0E022000DF91CF911F910F910895F894FFCFB9
+:0E022E0048656C6C6F20576F726C6421000085
+:00000001FF
