+++ alarm-uno/Debug/makedep.mk
@@ -1,0 +1,10 @@
+################################################################################
+# Automatically-generated file. Do not edit or delete the file
+################################################################################
+
+main.c
+
+password.c
+
+uart.c
+
