+++ alarm-uno/password.h
@@ -1,0 +1,5 @@
+#include <avr/io.h>
+#include <util/delay.h>
+
+int check_password(void);
+int compare(char *password, char *given_password);
