+++ arduino-alarm-project.atsln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Atmel Studio Solution File, Format Version 11.00
 VisualStudioVersion = 14.0.23107.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{54F91283-7BC4-4236-8FF9-10F437C3AD48}") = "arduino-alarm-project", "arduino-alarm-project\arduino-alarm-project.cproj", "{DCE6C7E3-EE26-4D79-826B-08594B9AD897}"
+Project("{54F91283-7BC4-4236-8FF9-10F437C3AD48}") = "alarm-uno", "alarm-uno\alarm-uno.cproj", "{74BBCDC1-1756-4FEE-9FB4-48D54763BF55}"
+EndProject
+Project("{54F91283-7BC4-4236-8FF9-10F437C3AD48}") = "alarm-mega", "alarm-mega\alarm-mega.cproj", "{2ECAD0FC-F307-4BE2-BB0E-AB67279DA0CF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +13,15 @@ Global
 		Release|AVR = Release|AVR
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{DCE6C7E3-EE26-4D79-826B-08594B9AD897}.Debug|AVR.ActiveCfg = Debug|AVR
-		{DCE6C7E3-EE26-4D79-826B-08594B9AD897}.Debug|AVR.Build.0 = Debug|AVR
-		{DCE6C7E3-EE26-4D79-826B-08594B9AD897}.Release|AVR.ActiveCfg = Release|AVR
-		{DCE6C7E3-EE26-4D79-826B-08594B9AD897}.Release|AVR.Build.0 = Release|AVR
+		{74BBCDC1-1756-4FEE-9FB4-48D54763BF55}.Debug|AVR.ActiveCfg = Debug|AVR
+		{74BBCDC1-1756-4FEE-9FB4-48D54763BF55}.Debug|AVR.Build.0 = Debug|AVR
+		{74BBCDC1-1756-4FEE-9FB4-48D54763BF55}.Debug|AVR.Deploy.0 = Debug|AVR
+		{74BBCDC1-1756-4FEE-9FB4-48D54763BF55}.Release|AVR.ActiveCfg = Release|AVR
+		{74BBCDC1-1756-4FEE-9FB4-48D54763BF55}.Release|AVR.Build.0 = Release|AVR
+		{2ECAD0FC-F307-4BE2-BB0E-AB67279DA0CF}.Debug|AVR.ActiveCfg = Debug|AVR
+		{2ECAD0FC-F307-4BE2-BB0E-AB67279DA0CF}.Debug|AVR.Build.0 = Debug|AVR
+		{2ECAD0FC-F307-4BE2-BB0E-AB67279DA0CF}.Release|AVR.ActiveCfg = Release|AVR
+		{2ECAD0FC-F307-4BE2-BB0E-AB67279DA0CF}.Release|AVR.Build.0 = Release|AVR
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
