--- arduino-alarm-project/Debug/arduino-alarm-project.hex
@@ -1,66 +1,0 @@
-:1000000071C0000086C0000084C0000082C00000F3
-:1000100080C000007EC000007CC000007AC00000EC
-:1000200078C0000076C0000074C0000072C00000FC
-:1000300070C000006EC000006CC000006AC000000C
-:1000400068C0000066C0000064C0000062C000001C
-:1000500060C000005EC000005CC000005AC000002C
-:1000600058C0000056C0000054C0000052C000003C
-:1000700050C000004EC000004CC000004AC000004C
-:1000800048C0000046C0000044C0000042C000005C
-:1000900040C000003EC000003CC000003AC000006C
-:1000A00038C0000036C0000034C0000032C000007C
-:1000B00030C000002EC000002CC000002AC000008C
-:1000C00028C0000026C0000024C0000022C000009C
-:1000D00020C000001EC000001CC000001AC00000AC
-:1000E00018C0000011241FBECFEFD1E2DEBFCDBF8C
-:1000F00000E00CBF12E0A0E0B2E0E8EFF3E000E0C7
-:100100000BBF02C007900D92AA30B107D9F702D0F9
-:1001100071C176CF26D0FFCF85B18E7F85B92FE311
-:100120008DE093E0215080409040E1F700C0000056
-:1001300085B1816085B92FE38DE093E02150804047
-:100140009040E1F700C0000085B18E7F85B92FE3B4
-:100150008DE093E0215080409040E1F700C0000026
-:10016000089584B1816084B984B1826084B984B116
-:10017000846084B984B1886084B98AB18F608AB997
-:1001800089EF90E00197F1F700C0000080EF8BB994
-:10019000C0E0D0E089B1803FE9F329B183EC99E078
-:1001A0000197F1F700C000009AB190959AB983ECDD
-:1001B00099E00197F1F700C000009BB190959BB9C1
-:1001C00083EC99E00197F1F700C0000089B1822B20
-:1001D00090E08737910571F481E3FE01E050FE4F16
-:1001E0008083FE01EB5FFD4F8081813309F0EEC01B
-:1001F00093DFECC08B37910571F482E3FE01E05090
-:10020000FE4F8083FE01EB5FFD4F8081823309F05A
-:10021000DDC082DFDBC08D37910571F483E3FE0121
-:10022000E050FE4F8083FE01EB5FFD4F8081833302
-:1002300009F0CCC071DFCAC08E37910571F481E43A
-:10024000FE01E050FE4F8083FE01EB5FFD4F808199
-:10025000813409F0BBC060DFB9C0873B910571F400
-:1002600084E3FE01E050FE4F8083FE01EB5FFD4F13
-:100270008081843309F0AAC04FDFA8C08B3B910571
-:1002800071F485E3FE01E050FE4F8083FE01EB5FD9
-:10029000FD4F8081853309F099C03EDF97C08D3BCB
-:1002A000910571F486E3FE01E050FE4F8083FE016C
-:1002B000EB5FFD4F8081863309F088C02DDF86C05B
-:1002C0008E3B910571F482E4FE01E050FE4F808385
-:1002D000FE01EB5FFD4F8081823409F077C01CDFA7
-:1002E00075C0873D910571F487E3FE01E050FE4F34
-:1002F0008083FE01EB5FFD4F8081873309F066C08C
-:100300000BDF64C08B3D910571F488E3FE01E05082
-:10031000FE4F8083FE01EB5FFD4F8081883309F043
-:1003200055C0FADE53C08D3D910571F489E3FE019D
-:10033000E050FE4F8083FE01EB5FFD4F80818933EB
-:1003400009F044C0E9DE42C08E3D910569F483E4C2
-:10035000FE01E050FE4F8083FE01EB5FFD4F808188
-:100360008334A1F5D9DE32C0873E910569F48AE273
-:10037000FE01E050FE4F8083FE01EB5FFD4F808168
-:100380008A3221F5C9DE22C08B3E910569F480E3F3
-:10039000FE01E050FE4F8083FE01EB5FFD4F808148
-:1003A0008033A1F4B9DE12C08D3E910579F08E3E06
-:1003B000910561F484E4FE01E050FE4F8083FE016C
-:1003C000EB5FFD4F8081843409F4A6DE8AB180950D
-:1003D0008AB989EF90E00197F1F700C000008BB176
-:1003E00080958BB9219687EA91E60197F1F700C0D5
-:0803F0000000D0CEF894FFCF0D
-:0A03F8007878787800303132330055
-:00000001FF
