--- arduino-alarm-project/Debug/arduino-alarm-project.srec
@@ -1,67 +1,0 @@
-S01D000061726475696E6F2D616C61726D2D70726F6A6563742E73726563B7
-S113000071C0000086C0000084C0000082C00000EF
-S113001080C000007EC000007CC000007AC00000E8
-S113002078C0000076C0000074C0000072C00000F8
-S113003070C000006EC000006CC000006AC0000008
-S113004068C0000066C0000064C0000062C0000018
-S113005060C000005EC000005CC000005AC0000028
-S113006058C0000056C0000054C0000052C0000038
-S113007050C000004EC000004CC000004AC0000048
-S113008048C0000046C0000044C0000042C0000058
-S113009040C000003EC000003CC000003AC0000068
-S11300A038C0000036C0000034C0000032C0000078
-S11300B030C000002EC000002CC000002AC0000088
-S11300C028C0000026C0000024C0000022C0000098
-S11300D020C000001EC000001CC000001AC00000A8
-S11300E018C0000011241FBECFEFD1E2DEBFCDBF88
-S11300F000E00CBF12E0A0E0B2E0E8EFF3E000E0C3
-S11301000BBF02C007900D92AA30B107D9F702D0F5
-S113011071C176CF26D0FFCF85B18E7F85B92FE30D
-S11301208DE093E0215080409040E1F700C0000052
-S113013085B1816085B92FE38DE093E02150804043
-S11301409040E1F700C0000085B18E7F85B92FE3B0
-S11301508DE093E0215080409040E1F700C0000022
-S1130160089584B1816084B984B1826084B984B112
-S1130170846084B984B1886084B98AB18F608AB993
-S113018089EF90E00197F1F700C0000080EF8BB990
-S1130190C0E0D0E089B1803FE9F329B183EC99E074
-S11301A00197F1F700C000009AB190959AB983ECD9
-S11301B099E00197F1F700C000009BB190959BB9BD
-S11301C083EC99E00197F1F700C0000089B1822B1C
-S11301D090E08737910571F481E3FE01E050FE4F12
-S11301E08083FE01EB5FFD4F8081813309F0EEC017
-S11301F093DFECC08B37910571F482E3FE01E0508C
-S1130200FE4F8083FE01EB5FFD4F8081823309F056
-S1130210DDC082DFDBC08D37910571F483E3FE011D
-S1130220E050FE4F8083FE01EB5FFD4F80818333FE
-S113023009F0CCC071DFCAC08E37910571F481E436
-S1130240FE01E050FE4F8083FE01EB5FFD4F808195
-S1130250813409F0BBC060DFB9C0873B910571F4FC
-S113026084E3FE01E050FE4F8083FE01EB5FFD4F0F
-S11302708081843309F0AAC04FDFA8C08B3B91056D
-S113028071F485E3FE01E050FE4F8083FE01EB5FD5
-S1130290FD4F8081853309F099C03EDF97C08D3BC7
-S11302A0910571F486E3FE01E050FE4F8083FE0168
-S11302B0EB5FFD4F8081863309F088C02DDF86C057
-S11302C08E3B910571F482E4FE01E050FE4F808381
-S11302D0FE01EB5FFD4F8081823409F077C01CDFA3
-S11302E075C0873D910571F487E3FE01E050FE4F30
-S11302F08083FE01EB5FFD4F8081873309F066C088
-S11303000BDF64C08B3D910571F488E3FE01E0507E
-S1130310FE4F8083FE01EB5FFD4F8081883309F03F
-S113032055C0FADE53C08D3D910571F489E3FE0199
-S1130330E050FE4F8083FE01EB5FFD4F80818933E7
-S113034009F044C0E9DE42C08E3D910569F483E4BE
-S1130350FE01E050FE4F8083FE01EB5FFD4F808184
-S11303608334A1F5D9DE32C0873E910569F48AE26F
-S1130370FE01E050FE4F8083FE01EB5FFD4F808164
-S11303808A3221F5C9DE22C08B3E910569F480E3EF
-S1130390FE01E050FE4F8083FE01EB5FFD4F808144
-S11303A08033A1F4B9DE12C08D3E910579F08E3E02
-S11303B0910561F484E4FE01E050FE4F8083FE0168
-S11303C0EB5FFD4F8081843409F4A6DE8AB1809509
-S11303D08AB989EF90E00197F1F700C000008BB172
-S11303E080958BB9219687EA91E60197F1F700C0D1
-S10B03F00000D0CEF894FFCF09
-S10D03F87878787800303132330051
-S9030000FC
